Extract common logic into `BaseDeploymentRecorder`
